feat: Separate Home component and mock data.
