+++ data/data.js
@@ -1,0 +1,35 @@
+export const data = {
+    avatar: 'https://avatars.githubusercontent.com/u/2822437?v=4',
+    message: 'This was done by a crazy developer',
+    repository: 'https://github.com/sandersgutierrez/static-website',
+    links: [
+        {
+            id: 1,
+            'name:': 'github',
+            url: 'https://github.com',
+            icon: 'fa-brands fa-github',
+            username: 'sandersgutierrez',
+        },
+        {
+            id: 2,
+            'name:': 'linkedin',
+            url: 'https://linkedin.com/in',
+            icon: 'fa-brands fa-linkedin',
+            username: 'sandersgutierrez',
+        },
+        {
+            id: 3,
+            'name:': 'twitter',
+            url: 'https://twitter.com',
+            icon: 'fa-brands fa-twitter',
+            username: 'untalsanders',
+        },
+        {
+            id: 4,
+            'name:': 'instagram',
+            url: 'https://instagram.com',
+            icon: 'fa-brands fa-instagram',
+            username: 'untalsanders',
+        },
+    ],
+}
